Update bitly_links
I don't have a small URL for this (yet), but wanted to test the code to see if it would pull the right information to where I expect it to go. Once I have this confirmed working, I'll request mzl.la URLs for the campaign ID referee links going forward for consistency.